🧼 Clean and Inform the code with other file
