+++ src/assets/scss/components/_login.scss
@@ -31,7 +31,11 @@
       flex-direction: column;
     }
   }
-  .login-right {
+  .fill {
+    min-height: 100vh;
+    height: 100%;
+  }
+  .wrapper {
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -39,44 +43,41 @@
     background-color: $red-magenta;
     border-top-left-radius: 91px;
     border-bottom-left-radius: 91px;
-  }
-  .fill {
-    min-height: 100vh;
-    height: 100%;
-  }
-  .form-group label:not(.chackboxLabel) {
-    display: inline-block;
-    font-size: medium;
-    color: $white;
-    padding: 10px;
-  }
-  input[type='email'],
-  input[type='password'] {
-    background: $white;
-    border: 1px solid $base-white;
-    border-radius: 26px;
-  }
-  .login-btn {
-    border-radius: 63px;
-    color: $black;
-    font-size: large;
-    font-weight: 700;
-    padding: 10px 60px;
-    border: 1px solid $baby-pink;
-    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-    cursor: pointer;
-    &:hover {
+    .form-group label:not(.chackboxLabel)  {
+      display: inline-block;
+      font-size: medium;
       color: $white;
-      background: $tint-black;
+      padding: 10px;
     }
-  }
-  .switch-page-btn {
-    margin-bottom: 75px;
-    margin-right: 20px;
+    input[type='email'],
+    input[type='password'] {
+      background: $white;
+      border: 1px solid $base-white;
+      border-radius: 26px;
+    }
+    .login-btn {
+      border-radius: 63px;
+      color: $black;
+      font-size: large;
+      font-weight: 700;
+      padding: 10px 60px;
+      border: 1px solid $baby-pink;
+      box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+      cursor: pointer;
+      &:hover {
+        color: $white;
+        background: $tint-black;
+      }
+    }
+    .switch-page-btn {
+      margin-bottom: 75px;
+      margin-right: 20px;
 
-    .current-page-tag {
-      &.active {
-        background: $white !important;
+      .current-page-tag {
+        &.active {
+          background: $white !important;
+          color: $red-magenta;
+        }
       }
     }
   }
